pyproject.toml
[tool.poetry]
name = "erp_backend"
version = "0.1.0"
description = "ERP Bakend"
authors = []
readme = "README.md"

[tool.poetry.dependencies]
django = "^4.2.5"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.0"
django-cities-light = "^3.9.2"
django-money = "^3.3.0"
django-phonenumber-field = { extras = ["phonenumbers"], version = "^7.1.0" }
drf-spectacular = "^0.26.4"
pillow = "^10.0.1"
psycopg2-binary = "^2.9.7"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.5.1"
python = "^3.8"
python-decouple = "^3.8"
dj-database-url = "^2.1.0"
gunicorn = "^21.2.0"
whitenoise = "^6.5.0"
jsonschema = "4.17.0"
colorama = "0.4.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ["py310"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "erp_backend.config.development"
python_files = ["tests/*.py"]
